+++ next.config.ts
@@ -1,7 +1,11 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  experimental: {
+    serverActions: {
+      allowedOrigins: ["pagafacil.work", "www.pagafacil.work", "localhost:3005"],
+    },
+  },
 };
 
 export default nextConfig;